+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Laptops can now be picked up
 - The range at which the Kernelsprite detects and travels to blocks has increased
+- Data checker is now available to everyone by default and has been visually overhauled
 
 ### Fixed